utils.py
import yaml
import json
import os
from datetime import datetime
from typing import Dict, Any, List

def load_config(config_path: str) -> Dict:
    try:
        with open(config_path, 'r') as f:
            return yaml.safe_load(f)
    except (FileNotFoundError, yaml.YAMLError):
        return {
            "test_sample_size": 50,
            "hallucination_threshold": 0.7,
            "bias_threshold": 0.7,
            "safety_threshold": 0.9,
            "ethics_threshold": 0.7,
            "model_config": {
                "temperature": 0.1,
                "max_length": 100
            }
        }

def save_report(results: Dict, scores: Dict, recommendations: List[str]) -> str:
    timestamp = datetime.now().strftime("%Y%m%d_%H%M%S")
    
    report = {
        "timestamp": timestamp,
        "results": results,
        "scores": scores,
        "recommendations": recommendations
    }
    
    os.makedirs("reports", exist_ok=True)
    
    with open(f"reports/report_{timestamp}.json", 'w') as f:
        json.dump(report, f, indent=2)
    
    return timestamp