requirements.txt
# Core Requirements
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
gradio>=3.36.0
sentence-transformers>=2.2.2

# Visualization & Utilities
matplotlib>=3.7.1
numpy>=1.24.3
pyyaml>=6.0
langchain-community>=0.0.11

# Optional (for full functionality)
huggingface_hub>=0.16.0
openai>=0.28.0